moya domashka_Python/domashka_21/requirements.txt
Pillow==10.0.0
pillow-heif==0.13.0
tqdm
